refactor: Paging interface 공통으로 변경
