+++ src/interface/index.ts
@@ -1,0 +1,6 @@
+export interface IPaging {
+  number: number;
+  size: number;
+  totalElement: number;
+  totalPages: number;
+}
